fix: Restructure resources for cleaner management
