+++ Dockerfile
@@ -17,6 +17,7 @@ RUN go mod download
 # Copy the go source
 COPY cmd/main.go cmd/main.go
 COPY api/ api/
+COPY internal/ internal/
 COPY pkg/ pkg/
 
 # Build
