+++ pkg/watcher/clusternimbuspolicy_watcher.go
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2023 Authors of Nimbus
+
 package watcher
 
 import (
